Fileの変換に関することを FilesConverter に集約
